requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
orjson==3.9.1
python-wootrade==0.1.10
requests==2.31.0
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
